fixed names to work with CMake Lists
